Refactor TestContext out of Testing.hpp
